move boxplot to separate file
